+++ src/components/FocusDisplay.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+function FocusDisplay() {
+  const mainFocus = localStorage.getItem("focus");
+  return <h2>Your Focus For Today is {mainFocus}</h2>;
+}
+
+export default FocusDisplay;
