--- src/context/TimeContext.js
@@ -1,5 +1,0 @@
-// import { createContext } from "react";
-
-// const TimeContext = createContext();
-
-// export default TimeContext;
