+++ src/pages/Focus.js
@@ -1,10 +1,12 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import Quote from "../components/Quote";
 import Clock from "../components/Clock";
 import FocusForm from "../components/FocusForm";
-import FocusItem from "../components/FocusItem";
+import FocusDisplay from "../components/FocusDisplay";
 
 function Focus() {
+  const [storageExist, setStorageExist] = useState(false);
+
   // const [focusItem, setFocusItem] = useState([{ id: null, task: "" }]);
   // const addTodo = (newTodoText) => {
   //   setFocusItem([...focusItem, { id: focusItem.id + 1, task: newTodoText }]);
@@ -20,9 +22,6 @@ function Focus() {
       <>
         <FocusForm />
       </>
-      {/* <>
-        <FocusItem />
-      </> */}
     </div>
   );
 }
